Cherry pick recent version of argpar (temporary)
Temporary commit to be able to use the most recent version of argpar.
The associated code will eventually be finalized and merged into master.

When rebasing on a more recent version of master with the merged updated
argpar the files below can be deleted and replaced with the new
versions.

Change-Id: Ifb688036664fcca591591a441d044b5078e7f09a
Signed-off-by: Erica Bugden <ebugden@efficios.com>