Add lttng-pause command
Pausing a recording session stops the consumption of tracing data to
the disk or network. The opposite command, lttng-resume restarts the
consumption of trace data.

The ability to pause and resume the consumption of tracing data
facilitates managing the total amount of trace data on a system at a
given time.

Since this command is a prototype:

• Pausing and resuming consumption of kernel trace data is
  unsupported. Currently, only userspace is supported. If the session is
  also tracing kernel space then consumption of that data will not be
  paused.

• Not all general options are implemented (e.g. machine interface) we
  just handle them without failing.

Change-Id: I01f378e8135ea9c7ee292f8d3f531e1159596dd3
Signed-off-by: Erica Bugden <ebugden@efficios.com>