Add lttng-resume command
Resuming a recording session restarts the consumption of tracing data
to the disk or network. The opposite command, lttng-pause stops the
consumption of trace data.

The ability to pause and resume the consumption of tracing data
facilitates managing the total amount of trace data on a system at a
given time.

Since this command is a prototype:

• Pausing and resuming consumption of kernel trace data is
  unsupported. Currently, only userspace is supported. If the session is
  also tracing kernel space then consumption of that data will not be
  paused.

• Not all general options are implemented (e.g. machine interface) we
  just handle them without failing.

Change-Id: I41a3364e35474cba0b7c9fe9c276d723540702aa
Signed-off-by: Erica Bugden <ebugden@efficios.com>